+++ example/rollup.js
@@ -967,6 +967,7 @@ function createNavigationTransition(context) {
     const destination = {
         url: entry.url,
         key: entry.key,
+        id: entry.id,
         index: destinationIndex,
         sameDocument: entry.sameDocument,
         getState() {
@@ -1307,7 +1308,13 @@ class Navigation extends NavigationEventTarget {
         }
         const nextUrl = new URL(url, baseURL).toString();
         let navigationType = "push";
-        if (options?.history === "push" || options?.history === "replace") {
+        if (options?.history === "auto" || !options?.history) {
+            // https://html.spec.whatwg.org/multipage/browsing-the-web.html#navigate-convert-to-replace
+            if (nextUrl === this.currentEntry?.url) {
+                navigationType = "replace";
+            }
+        }
+        else if (options?.history === "push" || options?.history === "replace") {
             navigationType = options?.history;
         }
         const entry = this.#createNavigationHistoryEntry({
@@ -1339,8 +1346,10 @@ class Navigation extends NavigationEventTarget {
         });
     };
     #createNavigationHistoryEntry = (options) => {
+        const key = options.key || (options.navigationType === "replace" ? this.currentEntry?.key : undefined);
         const entry = new NavigationHistoryEntry({
             ...options,
+            key,
             index: options.index ??
                 (() => {
                     return this.#entries.indexOf(entry);
@@ -1705,7 +1714,7 @@ class Navigation extends NavigationEventTarget {
     #dispose = async () => {
         // console.log(JSON.stringify({ known: [...this.#known], entries: this.#entries }));
         for (const known of this.#known) {
-            const index = this.#entries.findIndex((entry) => entry.key === known.key);
+            const index = this.#entries.findIndex((entry) => entry.id === known.id);
             if (index !== -1) {
                 // Still in use
                 continue;
@@ -2043,7 +2052,7 @@ async function transition(navigation) {
     while (navigation.transition && transition !== navigation.transition) {
         transition = navigation.transition;
         finalPromise = transition.finished;
-        await finalPromise.catch(error => void error);
+        await finalPromise.catch((error) => void error);
     }
     return finalPromise;
 }