+++ src/tests/custom-state.ts
@@ -5,49 +5,51 @@ import {ok} from "../is";
     const sessionStorageMap = new Map();
 
     const navigation = new Navigation({
-        getState({ key }) {
+        getState({ id }) {
             console.log("getState");
-            const maybe = sessionStorageMap.get(this.id);
+            const maybe = sessionStorageMap.get(id);
             if (!maybe) return undefined; // Or can return null here
             return maybe
         },
         setState(entry) {
             console.log("setState");
             const state = entry.getState();
-            const { key } = entry;
-            sessionStorageMap.set(key, state);
+            const { id } = entry;
+            sessionStorageMap.set(id, state);
         },
-        disposeState({ key }) {
+        disposeState({ id }) {
             console.log("disposeState");
-            sessionStorageMap.delete(key);
+            sessionStorageMap.delete(id);
         }
     })
 
     ok(!sessionStorageMap.size);
 
     const initialState = `Test ${Math.random()}`;
-    const { key: initialKey } = await navigation.navigate("/1", {
+    const { id: initialId } = await navigation.navigate("/1", {
         state: initialState
     }).finished;
 
     ok(sessionStorageMap.size);
-    const storedValue = sessionStorageMap.get(initialKey);
+    const storedValue = sessionStorageMap.get(initialId);
     ok(storedValue === initialState);
 
     const nextState = `Another ${Math.random()}`;
-    const { key: nextKey} = await navigation.navigate("/2", {
+    const { id: nextId, key: nextKey } = await navigation.navigate("/2", {
         state: nextState
     }).finished;
     ok(sessionStorageMap.size === 2);
-    const nextStoredValue = sessionStorageMap.get(nextKey);
+    const nextStoredValue = sessionStorageMap.get(nextId);
     ok(nextStoredValue === nextState);
 
-    const { key: finialKey } = await navigation.navigate("/2", {
+    const { id: finalId, key: finalKey } = await navigation.navigate("/2", {
         state: Math.random(),
         history: "replace"
     }).finished;
+    ok(nextKey === finalKey);
+    ok(nextId !== finalId);
     ok(sessionStorageMap.size === 2);
-    ok(sessionStorageMap.has(initialKey));
-    ok(!sessionStorageMap.has(nextKey));
-    ok(sessionStorageMap.has(finialKey));
+    ok(sessionStorageMap.has(initialId));
+    ok(!sessionStorageMap.has(nextId));
+    ok(sessionStorageMap.has(finalId));
 }